Make repos an associative array
There are uses of the repos array such as "${repos[$subcmd]}" and
"${repos[$cmd]}" which requires that the repos array be associative to
retrieve the correct value.

With an indexed array, the integer conversion of the subscript will be 0
which is not what is required. If the subscript contains a period ('.'),
bash returns an error trying to convert the subscript to an integer. For
example, trying to bash complete for 'add':

vcsh bashrc.d adbash: bashrc.d: syntax error: invalid arithmetic operator (error token is ".d")